// apps/web/src/lib/api/generated/models/timetableUpsertSlotBody.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Formex
 * OpenAPI spec version: 0.0.1
 */
import type { TimetableUpsertSlotBodySubjectsItem } from "./timetableUpsertSlotBodySubjectsItem";

export type TimetableUpsertSlotBody = {
  /**
   * @minimum 1
   * @maximum 12
   */
  semester: number;
  day: string;
  /**
   * @minimum 1
   * @maximum 7
   */
  start_hour_no: number;
  /**
   * @minimum 1
   * @maximum 7
   */
  end_hour_no: number;
  /**
   * @minItems 1
   * @maxItems 10
   */
  subjects: TimetableUpsertSlotBodySubjectsItem[];
};